Fixed a bug in the s2Member gateway.
A capability that s2Member does not remove at EOT was granted to users.
This caused users to be able to illegitimately view protected content.

Also when no current EOT was set, a user would purchase a subscription
that ended the day after it was bought, regardless of what time the
subscription should last.