Tested up to 5.2.
Hello, Remco Tolsma!

WordPress 5.2 is scheduled to be released on May 7. Are your plugins ready?

After testing your plugins and ensuring compatibility, it only takes a few moments to change the readme "Tested up to:" value to 5.2. This information provides peace of mind to users and encourages them to update to the latest version.

Here are the current "Tested up to:" values for each of your plugins:

 * https://wordpress.org/plugins/pronamic-ideal/ (tested up to 5.1.1)

For each plugin that is compatible, you don't need to release a new version — just change the stable version's readme value.

Looking to get more familiar with 5.2? Read this roundup post on the core development blog to learn more about the new fatal error recovery mode, new network and sites query filters, the new Block Editor JavaScript module, new Dashicons, updates to admin tab semantics, and updates to accessibility and privacy:

https://make.wordpress.org/core/2019/04/16/wordpress-5-2-field-guide/

There are three areas that specific focus should be given:

1.) Note the shim provided for themes wanting to utilize the wp_body_open() function in older versions of WordPress and prevent fatal errors from the undefined function. Read the Miscellaneous Developer Updates dev note for details on the function and shim:
https://make.wordpress.org/core/2019/04/24/miscellaneous-developer-updates-in-5-2/

2.) Note the shim provided for plugins and themes wanting to support the new is_privacy_policy() function in older versions of WordPress. Read the Privacy dev note for details on the function and shim:
https://make.wordpress.org/core/2019/04/24/developer-focused-privacy-updates-in-5-2/

3.) The wp-includes/fonts/dashicons.woff file does NOT include the 13 new icons. Read the Dashicons dev note for information on options available:
https://make.wordpress.org/core/2019/04/11/dashicons-in-wordpress-5-2/

Thank you for all you do for the WordPress community, and we hope you enjoy 5.2 as much as we do.

WordPress core contributors